iconApp-main/requirement.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
dataclasses==0.8
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-countries==7.3.2
django-credit-cards==0.4.1
django-dotenv==1.4.2
django-filter==21.1
django-model-utils==4.2.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Flask==2.0.3
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.2.0
phonenumbers==8.12.46
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
